### 2021-12-21 #4018 @whedon
[PRE REVIEW]: OfflineMOT: A Python Package for multiple objects detection and tracking from bird view stationary drone videos
**Submitting author:** @engyasin (<a href="http://orcid.org/0000-0002-5282-7259">Yasin Yousif</a>)
**Repository:** <a href="https://github.com/engyasin/Offline_MOT" target ="_blank">https://github.com/engyasin/Offline_MOT</a>
**Version:** v1.0.0
