### 2021-12-29 #4018 @arfon
@engyasin – many thanks for your submission to JOSS. While I look for a possible editor, if you want to fix the DOIs (noting that Whedon's suggestions are not always right), you can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
