+++ paper/paper.bib
@@ -16,6 +16,7 @@
   author={Bergmann, Philipp and Meinhardt, Tim and Leal-Taixe, Laura},
   booktitle={Proceedings of the IEEE/CVF International Conference on Computer Vision},
   pages={941--951},
+  doi={10.1109/iccv.2019.00103},
   year={2019}
 }
 
@@ -25,6 +26,7 @@
   booktitle={2017 IEEE international conference on image processing (ICIP)},
   pages={3645--3649},
   year={2017},
+  doi={10.1109/icip.2017.8296962},
   organization={IEEE}
 }
 
@@ -116,5 +118,6 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
   booktitle={2019 IEEE Intelligent Vehicles Symposium (IV)},
   pages={899--904},
   year={2019},
+  doi={10.1109/ivs.2019.8814092},
   organization={IEEE}
 }
