### 2022-01-26 #4099 @whedon
[REVIEW]: OfflineMOT: A Python Package for multiple objects detection and tracking from bird view stationary drone videos
**Submitting author:** @engyasin (<a href="http://orcid.org/0000-0002-5282-7259">Yasin Yousif</a>)
**Repository:** https://github.com/engyasin/Offline_MOT
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.9
