### 2022-02-15 #4099 @engyasin
> I am reviewing...
>
> I found several problems:
>
> * a lot of typo erros in the **README.md**, but didn't affect the understanding
> * I tried to run the first command `python offline_mot\main.py -v [directroy of the videos, ex: docs\sample.mp4]`, but failed.
>
> @hugoledoux , what should I do next? Should I propose an issur in the original repo?

Thanks for the feedback,

* For the first point, I will fix the typos
* For the second point, can you show the error (whether in a new issue or here) it could be the result of missing library, different version, etc.
