### 2022-02-15 #4099 @Shiming94
Hi, @engyasin, the Traceback is shown blow:
```
Traceback (most recent call last):
  File "offlinemot/main.py", line 401, in <module>
    main(args)
  File "offlinemot/main.py", line 184, in main
    objects,detector = FirstFrame(frame)
  File "offlinemot/main.py", line 147, in FirstFrame
    detector  = YoloDetector(config.model_config,config.model_name,use_cuda=config.use_cuda)
  File "/home/shiming/12.Offline_MOT/offlinemot/detection.py", line 61, in __init__
    self.m.load_state_dict(torch.load(weightfile))
  File "/home/shiming/anaconda3/envs/mot/lib/python3.7/site-packages/torch/serialization.py", line 608, in load
    return _legacy_load(opened_file, map_location, pickle_module, **pickle_load_args)
  File "/home/shiming/anaconda3/envs/mot/lib/python3.7/site-packages/torch/serialization.py", line 777, in _legacy_load
    magic_number = pickle_module.load(f, **pickle_load_args)
_pickle.UnpicklingError: invalid load key, 'v'.
```
I think it should be some error in the config file.
