+++ outputs/sample.txt
@@ -1,3 +1,4 @@
+
 101 [1848, 496, 35, 53] 1 19 -169
 102 [1849, 496, 35, 53] 1 19 -168
 103 [1848, 496, 35, 53] 1 19 -169
