+++ Readme.md
@@ -2,16 +2,16 @@
 Multiple objects detection and tracking from bird view stationary drone videos
 =========
 
-This project is used to extract trajectories as precise as possible for different entites in a mixed traffic scene, from drone videos taken while hovering in the same position (stationary).
+This project is used to extract trajectories as precise as possible for different entities in a mixed traffic scene, from drone videos taken while hovering in the same position (stationary).
 
 The project needs a pretrained network for detection, which should be trained separately. The network included with this library is Yolov4 in a pytorch format. The loading and running of Yolo model is done with the help of scripts taken from [this project](https://github.com/Tianxiaomo/pytorch-YOLOv4) (All of them are in *offlinemot/tool* subfolder)
 
-Example output for a sample video, taken from [**Vehicle-Crowd Intraction (VCI) - CITR Dataset**](https://github.com/dongfang-steven-yang/vci-dataset-citr) :
+Example output for a sample video, taken from [**Vehicle-Crowd Interaction  (VCI) - CITR Dataset**](https://github.com/dongfang-steven-yang/vci-dataset-citr) :
 
 ![Problem loading the gif!](docs/output.gif)
 
 
-***This example shows some minor problems because the scene, the backgorund and the drone's camera are outside the detection network training set (never seen before by the detection network).***
+***This example shows some minor problems because the scene, the background and the drone's camera are outside the detection network training set (never seen before by the detection network).***
 
 ***However, the application of this project (including the Yolo network training) was targeted for a Cyclists dataset videos [to be cited later], where all the trajectories where extracted successfully.***
 
@@ -20,21 +20,21 @@ Example output for a sample video, taken from [**Vehicle-Crowd Intraction (VCI) 
 
 Drone videos are subject to some random movements due to a few factors like noise in the control or wind. For that a fixing step for all the affected frames is needed. 
 
-Additionally running the detection for every frame is slow and can be substitued by the usage of a background substarction method to detect all moving objects. If the video are stationary (like the assumption here) then the background is the same for the whole video.
+Additionally running the detection for every frame is slow and can be substituted by the usage of a background subtracting method to detect all moving objects. If the video are stationary (like the assumption here) then the background is the same for the whole video.
 
 At the end and because the work is offline, additional filtering steps can be done to find: 
 
 - The true size of each object (which should be the same because it is taken from bird's eye view)
 - The orientation of all the objects in each frame
-- Solving the errors of misclassifying some objects by assigning the highst and most probable detected class to them. 
+- Solving the errors of misclassifying some objects by assigning the highest and most probable detected class to them. 
 - Smoothing the trajectories.
 
-All this is feasable because of the specific features of the problem, namely (bird's eye view, offline and 
+All this is feasible because of the specific features of the problem, namely (bird's eye view, offline and 
 stationary camera)
 
 ## Getting Started
 
-After cloning this project, and changing the detection model (in case you need to detect other objects than cyclists, pedistrains and cars), the requirerments packages should be installed, so simply, cd to the root of this project and run:
+After cloning this project, and changing the detection model (in case you need to detect other objects than cyclists, pedestrians and cars), the requirements packages should be installed, so simply, cd to the root of this project and run:
 
 ```
 pip install -r requirements.txt
@@ -44,28 +44,28 @@ pip install -r requirements.txt
 
 There are many parameters that should be tuned before running the tracking on a new video. All of them are in the `offline_mot/config.py` file. 
 
-A detailed explaination for all of them is avaliable on the same file above as well as in the docs folder, namely [Working example](./docs/A_Working_Example.ipynb)
[…]
