### 2022-02-15 #4099 @engyasin
Thanks for reporting @Shiming94

I tried to reproduce the error, but I couldn't.
It seems it is related to the Yolo network file (models/Yolo4_epoch300.pth). it is a big file (244 MB), which needs git lfs to be uploaded. Can you check if it's downloaded correctly in your cloned repo?
