### 2022-02-15 #4099 @Shiming94
@engyasin I just downloaded all the files in your repo and followed your instructions, and then the error appeared. I will try the new checkpoint file like you said. But I recommend that you create a download script at the beginning of your Readme.md. That will make the followers and the audience much easier to follow.
