### 2022-02-15 #4099 @engyasin
Yes, that's right,
I will test it on another device without git lfs. and update the readme with the needed instructions.
