### 2022-02-15 #4099 @lucaferranti
Hi @engyasin , overall very nice work, here are some preliminary comments regarding the software

- [ ] **Installation**: Currently the software is installed by cloning the repository and doing the installation manually. Have you considered adding this to pypi? This way I could install it with just `pip install offlinemot`. In the example notebooks, you have indeed a first cell to manually add the path of `offlinemot` and be able to import functionality. This can be a little annoying for other people using the software in their research. It also makes sharing work harder, imagine I write a notebook that uses your software and share it with a colleague, it would require my colleague to clone your repository and adjust the path to match their installation.
- [ ] **Documentation 1**: With a quick glance at the source code, most functions seem have excellent docstrings. Unfortunately, these docstrings are not collected into a full API documentation. This makes it hard for the user to investigate the interface of the package. I would strongly suggest to generate a documentation page for the software, e.g. using sphinx. The documentation page should have the API documentation and could also have the current notebooks as tutorials.
- [ ] **Testing**: I get some tests failing when testing locally, possibly for the reason reported by the other reviewer. I will dig deeper into this and open an issue if needed. Anyway, I think the current setup where tests are run locally can make maintaining the package harder, mainly for two reasons. First, if someone opens a PR to your code, either you trust they have already tested locally or you have to clone their branch and test the tests locally every time. Second, unless you have several different machines, tests will be run only on one operating system, hence OS dependent failures are easier to slip through (the most common case would be that things work on linux/macos but not on windows, these bugs are fairly usual if you work with paths). For this reason, I would strongly suggest to setup a CI workflow (e.g with github actions) to run the tests on each push and PR, possibly on different operating systems.
- [ ] **Contributing**: I feel the current "How to contribute" section doesn't say how to contribute, as it merely encourages to read the documentation and browse the code. I would suggest some rewording, for example just replicating there the first two sentences of the Support section in the readme would make it much better. Style guide is missing, but fwiw I don't think this is strictly needed to be done before getting the paper published (but please do add that at some point :) )
- [ ] **Functionality**: I get the same error reported by the other reviewer, a couple of more comments
   - The example command in the readme has probably a typo, I think should be `offline_mot/main.py` -> `offlinemot/main.py`, same for the other examples
   - That command works only assuming you are in the right place, this should be specified. Again, I think ideally I should be able to install the package and run it from anywhere
   - Can I use the package with other detection models? Do they have to be theano or can it be a tensorflow model? It would be nice to have an example about this.

 This was the first round about the code, I'll now read the paper.
