### 2022-02-15 #4099 @lucaferranti
regarding the issue reported by @Shiming94 , I checked my cloned repository and the `model` folder seems to match the content on this repository, so I think it did download the network?
```
lferrant@LAPTOP-HGM2FD2S:/mnt/c/Users/lucaa/Desktop/Offline_MOT/model$ ll
total 1356
drwxrwxrwx 1 lferrant lferrant     512 Feb 15 13:07 ./
drwxrwxrwx 1 lferrant lferrant     512 Feb 15 20:00 ../
-rwxrwxrwx 1 lferrant lferrant 1373697 Feb 15 13:07 00120.jpg*
-rwxrwxrwx 1 lferrant lferrant     134 Feb 15 13:07 Yolov4_epoch300.pth*
-rwxrwxrwx 1 lferrant lferrant      22 Feb 15 13:07 obj.names*
-rwxrwxrwx 1 lferrant lferrant   12248 Feb 15 13:07 yolov4-obj.cfg*
-rwxrwxrwx 1 lferrant lferrant     134 Feb 15 13:07 yolov4-pytorch.pth*
```

anyway, I am not sure such big data files belong to the source code (but you can disagree). The readme states that a separated model for detection is needed, so I feel the users should be able to install the package without having to download such a heavy model (e.g. if they want to use a different one anyway in their research project). An alternative could be to have a smaller lightweight model to make the package "battery included" and have a script/function to download the yolo4 network if the user wants to use it.

When running the tests, you could check if the yolo4 network is downloaded in a predesigned location and if it isn't download it. This way you could still run the tests on the CI and you wouldn't redownload the big file every time you run tests locally. A maybe even better alternative could be to run the tests with the smaller lightweight model, the final performance metrics will probably be lower than with yolo, but I think that's orthogonal to testing for correctness.
