### 2022-02-20 e9c4ed8
docs and pypi setup
