+++ .github/workflows/pytest.yml
@@ -15,7 +15,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v2
         with:
-          python-version: [3.7, 3.8]
+          python-version: ["3.7", "3.8"]
 
       - name: Install dependencies
         run: |
