+++ docs/build/html/_sources/generated/offlinemot.utils.save_tracks.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.utils.save\_tracks
-=============================
-
-.. currentmodule:: offlinemot.utils
-
-.. autofunction:: save_tracks
