+++ docs/build/html/_sources/generated/offlinemot.utils.check_box.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.utils.check\_box
-===========================
-
-.. currentmodule:: offlinemot.utils
-
-.. autofunction:: check_box
