+++ docs/build/html/_sources/generated/offlinemot.utils.read_tracks.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.utils.read\_tracks
-=============================
-
-.. currentmodule:: offlinemot.utils
-
-.. autofunction:: read_tracks
