+++ docs/build/html/_sources/generated/offlinemot.utils.find_overlap.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.utils.find\_overlap
-==============================
-
-.. currentmodule:: offlinemot.utils
-
-.. autofunction:: find_overlap
