+++ docs/build/html/_sources/generated/offlinemot.utils.detect_overlaping.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.utils.detect\_overlaping
-===================================
-
-.. currentmodule:: offlinemot.utils
-
-.. autofunction:: detect_overlaping
