+++ docs/build/html/_sources/generated/offlinemot.utils.rst.txt
@@ -1,32 +1,0 @@
-﻿offlinemot.utils
-================
-
-.. rubric:: Description
-
-.. automodule:: offlinemot.utils
-
-.. currentmodule:: offlinemot.utils
-
-
-
-
-.. rubric:: Functions
-
-.. autosummary::
-    :toctree: .
-    
-    check_box
-    
-    detect_overlaping
-    
-    find_overlap
-    
-    read_tracks
-    
-    resize
-    
-    save_tracks
-    
-    transform_detection
-    
-
