+++ setup.py
@@ -22,9 +22,9 @@ setup(
   license='MIT',        # Chose a license from here: https://help.github.com/articles/licensing-a-repository
   description = short_description[0],   # Give a short description about your library
   author = 'Yasin Yousif',                   # Type in your name
-  author_email = 'yasin.yousif@tu-clausthal',      # Type in your E-Mail
+  author_email = 'mrhidestev@hotmail.com',      # Type in your E-Mail
   url = 'https://github.com/engyasin',   # Provide either the link to your github or to your website
-  download_url = 'https://github.com/engyasin/Offline_MOT/archive/v_01.tar.gz',    # I explain this later on
+  download_url = 'https://github.com/engyasin/Offline_MOT/archive/refs/tags/v1.0.0.tar.gz',    # I explain this later on
   keywords = ['Multi Objects Tracking', 'Detection', 'Traffic Data'],   # Keywords that define your package best
   install_requires=[            # I get to this in a second
           'numpy',
