+++ offlinemot/model/sample.mp4.txt
@@ -1,7 +1,17 @@
-3 [1752, 513, 63, 36] 1 0 165
-4 [1753, 512, 63, 36] 1 0 164
-5 [1753, 512, 63, 36] 1 0 164
-6 [1750, 513, 63, 36] 1 0 164
-7 [1747, 514, 63, 36] 1 0 164
-8 [1743, 515, 63, 36] 1 0 164
-9 [1739, 515, 63, 36] 1 0 164
+3 [1752, 509, 62, 44] 1 0 166
+4 [1754, 508, 62, 44] 1 0 164
+5 [1753, 508, 62, 44] 1 0 165
+6 [1751, 509, 62, 44] 1 0 167
+7 [1747, 510, 62, 44] 1 0 171
+8 [1743, 511, 62, 44] 1 0 179
+9 [1741, 511, 62, 44] 1 0 -171
+10 [1740, 512, 62, 44] 1 0 -171
+11 [1738, 513, 62, 44] 1 0 -171
+12 [1735, 513, 62, 44] 1 0 -171
+13 [1733, 514, 62, 44] 1 0 -171
+14 [1732, 514, 62, 44] 1 0 -171
+15 [1732, 513, 62, 44] 1 0 -171
+16 [1732, 513, 62, 44] 1 0 -171
+17 [1731, 512, 62, 44] 1 0 -171
+18 [1729, 511, 62, 44] 1 0 -171
+19 [1728, 509, 62, 44] 1 0 -171
