+++ .github/workflows/pytest.yml
@@ -4,8 +4,8 @@ on: [push,pull_request]
 
 jobs:
   test:
-    runs-on: [macOS-latest, ubuntu-latest]
-    timeout-minutes: 10
+    runs-on: ubuntu-latest
+    timeout-minutes: 20
 
     steps:
       - name: Check out repository code
