+++ .github/workflows/pytest.yml
@@ -21,6 +21,7 @@ jobs:
         run: |
           python3 -m pip install --upgrade pip
           pip3 install -r requirements.txt
+          python3 -c "import offlinemot"
       - name: Run test suite
         run: |
-          pytest --exitfirst --verbose --failed-first 
+          pytest offlinemot/tests --exitfirst --verbose --failed-first 
