### 2022-02-21 #4099 @lucaferranti
and here the second part of my review for the paper
- [ ] **Authorship**: could you clarify the contribution of Awad Mukbil and Jörg Müller to the package development? Based on the list of commits, they don't seem to be contributors

- [ ] **State of the field**: The paper at the moment compares the presented method to other methods, since this is a software paper, I think the focus should be on comparing the implementation to other existing packages. Are there other packages / codes tackling this problem, how does offlineMOT compare to them? At the moment, the section tells me some approaches have been developed but I am missing pointers to if and where they are implemented.
- [ ] **Example usage**: I find this section in the current form a little confusing, I don't think installation instructions are needed in the paper, they should be covered in the documentation / readme. Furthermore, the current section is briefly showing a command line to call a specific script and commenting the results. If some usage examples are to be included, I think they should be more focused on the interface of the software and not the result. For example, before the pseudocode it is mentioned that the package is implemented in a modular way and it's easy to turn off and/or tune different parts, instead of the pseudocode, I think it would be more useful to have a code snippet showing how to build each block (if this can be done with a few lines of code, if the example is longer, better to just refer to the documentation)
