### 2022-02-22 #4099 @hugoledoux
> * a lot of typo erros in the **README.md**, but didn't affect the understanding

@Shiming94 if there are many typos then the authors should go over the text and indeed fix them. If you spot a few, the best would be to fork the paper, fix them, and make a pull request.
