### 2022-02-22 #4099 @hugoledoux
> anyway, I am not sure such big data files belong to the source code (but you can disagree). The readme states that a separated model for detection is needed, so I feel the users should be able to install the package without having to download such a heavy model (e.g. if they want to use a different one anyway in their research project). An alternative could be to have a smaller lightweight model to make the package "battery included" and have a script/function to download the yolo4 network if the user wants to use it.

@engyasin I share the opinion of @lucaferranti here, ideally your main test dataset should be small, small enough to be committed to a standard git repo at least. Most users just want to `pip install X` (so I think you should consider this too, although not a strict requirement for JOSS) and then test if the code works
