### 2022-02-22 #4099 @hugoledoux
> * [ ]  **Authorship**: could you clarify the contribution of Awad Mukbil and Jörg Müller to the package development? Based on the list of commits, they don't seem to be contributors

@engyasin please read this: https://joss.readthedocs.io/en/latest/submitting.html?highlight=contributors#authorship JOSS is slightly different from mainstream academic publishers and has rules for authorship. Of course it's not black-and-white, so you can just explain us here what was done by those who didn't contribute to the code (in the git commits)
