+++ paper/paper.md
@@ -25,25 +25,25 @@ bibliography: paper.bib
 # Summary
 
 The topic of multi objects tracking (MOT) is still considered an open research area [@MOTChallenge20].
-Among the many available methods for this problem, it is worth mentioning *Deep Sort* [@wojke2017simple], where a detection and tracking steps is done in real time and for different types of scenes and areas (for example on pedestrian movement or for vehicles tracking). Another state-of-the-art method is *Tracktor* [@bergmann2019tracking], where tracking is done by repetitive detections on all the frames in the video. 
+Among the many available methods for this problem, it is worth mentioning *Deep Sort* [@wojke2017simple], where a detection and tracking steps are done in real time and for different types of scenes and areas (for example for pedestrians movement or for vehicles tracking). Another state-of-the-art method is *Tracktor* [@bergmann2019tracking], where tracking is done by repetitive detections on all the frames in the video. 
 
 The importance of the problem comes from its many applications, for example self-driving cars software, traffic analysis, or general surveillance applications.  
-Unfortunately, due to the variety of scenes and contexts and due to the time constrains that are needed for some applications, there is no one general solution that is able to work perfectly for all cases. 
+Unfortunately, due to the variety of scenes and contexts and due to the time constrains that are needed for some applications, there is no one general solution capable of working perfectly for all cases. 
 
-For example, for the two cases of a moving camera recording side view of a road traffic, and a drone camera recording from above, there are many different challanges that should be addressed for each case. Developing one method for both cases, will make this method less effective for addressing each case problems alone. 
+For example, between the two cases of a moving camera recording side view road traffic, and a drone camera recording from above, there are many different challanges that should be addressed for each case. Developing one method for both cases, will make this method less effective for addressing each case problems alone. 
 
- `OfflineMOT`, the package introduced here, tries to provide a solution to a more restricted problem, namely bird’s eye stationary videos without real-time condition. It applies mainly three techniques for detection and tracking on three different priority levels.
+ `OfflineMOT`, the package introduced here, tries to provide a solution to a more restricted problem, namely bird’s eye stationary videos without real-time condition. It applies mainly three techniques for detection and tracking on three different priority levels. Below they are listed from lowest to highest.
 
-- The first level is **background subtraction**. It is a fast method to find what pixels have changed in the image (the forground), and this is plausable because of the stationary condition. Otherwise, if the drone is moving freely, then the background will be less learnable. 
-Another problem here is the subtle movement of the drone due to wind and control noise. To solve this, a program for fixing the view is implemented. It depends on matching the background features with every next frame, and then transforming the frame if big movement is detected.
+- The first level is **background subtraction**. It is a fast method to find what pixels have changed in the image (the forground), and this is possible because of the stationary condition. Otherwise, if the drone's camera is moving freely, then the background will be less learnable. 
+Another problem here is the subtle movement of the drone due to wind and control noise. To solve this, a program for fixing the view is implemented. It is based on matching the background features with every next frame, and then transforming the frame if a big movement is detected.
 
-- The second level is **multi-objects tracking** methods such as *kernelized correlation filters* (KCF) [@henriques2014high]. This method takes the output of the detection with the next frame and it finds where the objects will be. It can also return failure or success states of tracking.
+- The second level is **multi-objects tracking** methods such as *kernelized correlation filters* (KCF) [@henriques2014high]. This method takes the output of the detection and the next frame and it gives the next postions of the objects. It can also return the failure or success states of tracking.
 
[…]
