+++ setup.py
@@ -18,13 +18,13 @@ except:
 setup(
   name = 'offlinemot',         # How you named your package folder (MyLib)
   packages = find_packages(),   # Chose the same as "name"
-  version = '1.0.0',      # Start with a small number and increase it with every change you make
+  version = '1.0.1',      # Start with a small number and increase it with every change you make
   license='MIT',        # Chose a license from here: https://help.github.com/articles/licensing-a-repository
   description = short_description[0],   # Give a short description about your library
   author = 'Yasin Yousif',                   # Type in your name
   author_email = 'mrhidestev@hotmail.com',      # Type in your E-Mail
   url = 'https://github.com/engyasin',   # Provide either the link to your github or to your website
-  download_url = 'https://github.com/engyasin/Offline_MOT/archive/refs/tags/v1.0.0.tar.gz',    # I explain this later on
+  download_url = 'https://github.com/engyasin/Offline_MOT/archive/refs/tags/v1.0.1.tar.gz',   # I explain this later on
   keywords = ['Multi Objects Tracking', 'Detection', 'Traffic Data'],   # Keywords that define your package best
   install_requires=[            # I get to this in a second
           'numpy',
