+++ docs/source/conf.py
@@ -68,6 +68,13 @@ exclude_patterns = ['_build', '**.ipynb_checkpoints']
 #
 html_theme = 'sphinx_rtd_theme'
 
+html_context = {
+  'display_github': True,
+  'github_user': 'engyasin',
+  'github_repo': 'Offline_MOT',
+  'github_version': 'main/docs/source/',
+}
+
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
