+++ offlinemot/tests/example_data.py
@@ -38,9 +38,9 @@ results_2 = [[(100,120),(200,220),0.9,-1],
             [(7,120),(105,157),0.8,-1]]
 
 # boxes for detection
-detect_boxes = [[100,50,100,100],
-                [0,0,100,150],
-                [7,50,98,37]]
+detect_boxes = [(100,50,100,100),
+                (0,0,100,150),
+                (7,50,98,37)]
 
 # tracks and angels
 
