+++ paper/paper.md
@@ -67,97 +67,6 @@ offlinemot.main.set_params()
 In the following table, the most important ones of these parameters are listed along with how to tune them for a new video. 
 
 ![Important parameters to tune in config.py \label{table:parameters}](table.PNG)
-<!-->
-<html>
-</br>
-</br>
-</br>
-    <table align="center" border="1">
-        <caption> Table 1: Important parameters to tune in <code>config.py</code></caption>
-        <thead>
-            <tr>
-                <th>Tracking and </br>Detection</br> Module</th>
-                <th>The Parameter</th>
-                <th>The Function</th>
-            </tr>
-        </thead>
-        <tbody>
-            <tr>
-                <td rowspan=3>General Settings</td>
-                <td> <code> detect_every_N </code> </td>
-                <td> The frame rate in which the detection is performed. </br> 
-                    Higher values make the processing faster but mistakes become more probable. For optimal accuracy: 1</td>
-            </tr>
-            <tr>
-                <td> <code>missing_thresh</code></td>
-                <td> The threshold for deleting a tracked object if its fail-to-success tracking ratio is lower than this number. </br>
-                If the objects are deleted fast consider decreasing this number</td>
-            </tr>
-            <tr>
-                <td> <code>colors_map</code> </td>
-                <td>It is a list of RGB colors to annotate tracked objects based on class. </br>
-                Define your object classes ids and set a color for each of them there starting from object: 1</td>
-            </tr>
-            <tr>
-                <td rowspan=2>Background </br> Subtraction</td>
-                <td> <code>bgs_shadows</code></td>
-                <td> Set to <code>True</code> in case your video has strong shadows.</td>
-            </tr>
-            <tr>
-                <td> <code>bgs_min_area</code> </td>
-                <td> Set to 90% of the smallest object area to be tracked.</td>
-            </tr>
-            <tr>
-                <td rowspan=1>View Fixing</td>
-                <td> <code>do_fix</code> </td>
-                <td> If your video has some fast moves that wil disturb the tracking, then set to True</td>
-            </tr>
-            <tr>
-                <td rowspan=4>Detection </td>
-                <td>
-                        <code>model_name</code>  </br>    
-                        <code>model_config</code>     </br>
-                       <code>classes_file_name</code> 
-                </td>
-                <td>These three parameters are to be loaded when you train Yolo on your own dataset.
-                </br>
-                They are respectively for the model file path, configuration file path and classes names file path. 
-                </td>
-            </tr>
-            <tr>
-                <td> <code>detection_thresh</code> </td>
-                <td> It is the minimum threshold for the detection probability. 
-                </br>
-                If you want more detections but with higher false positive rate, you may set it to lower values 
-                </td>
-            </tr>            
-            <tr>
-                <td> <code>dist_thresh</code> </td>
-                <td> The distance threshold to assign a detection to a tracked object</td>
-            </tr>            
-            <tr>
-                <td> <code>size_thresh</code> </td>
-                <td> The minimum threshold of size difference between a detection and assigned tracked object</td>
-            </tr>
-            <tr>
-                <td rowspan=3>Filtering and </br> Smoothing</td>
-                <td> <code>overlap_thresh</code> </td>
-                <td> The maximum intersection to object's area ratio between two objects to delete the first object </td>
-            </tr>
-            <tr>
-                <td> <code>do_smooth</code> </td>
-                <td>A Boolean to determine whether to smooth trajectories or not</td>
-            </tr>
-            <tr>
-                <td> <code>save_out_video</code> </td>
-                <td> Whether to save the output annotated video or not</td>
-            </tr>
-        </tbody>
-    </table>
-
-</html>
-
-<-->
 
 # Scope
 
