### 2022-02-27 #4099 @engyasin
Hello,
In the following, the responses to previously raised issues regarding the submission are listed

> * [ ]  **Installation**: Currently the software is installed by cloning the repository and doing the installation manually. Have you considered adding this to pypi? This way I could install it with just `pip install offlinemot`. In the example notebooks, you have indeed a first cell to manually add the path of `offlinemot` and be able to import functionality. This can be a little annoying for other people using the software in their research. It also makes sharing work harder, imagine I write a notebook that uses your software and share it with a colleague, it would require my colleague to clone your repository and adjust the path to match their installation.

The package is now on pypi.

> * [ ]  **Documentation 1**: With a quick glance at the source code, most functions seem have excellent docstrings. Unfortunately, these docstrings are not collected into a full API documentation. This makes it hard for the user to investigate the interface of the package. I would strongly suggest to generate a documentation page for the software, e.g. using sphinx. The documentation page should have the API documentation and could also have the current notebooks as tutorials.

Documentation with full API is in https://engyasin.github.io/Offline_MOT

> * [ ]  **Testing**: I get some tests failing when testing locally, possibly for the reason reported by the other reviewer. I will dig deeper into this and open an issue if needed. Anyway, I think the current setup where tests are run locally can make maintaining the package harder, mainly for two reasons. First, if someone opens a PR to your code, either you trust they have already tested locally or you have to clone their branch and test the tests locally every time. Second, unless you have several different machines, tests will be run only on one operating system, hence OS dependent failures are easier to slip through (the most common case would be that things work on linux/macos but not on windows, these bugs are fairly usual if you work with paths). For this reason, I would strongly suggest to setup a CI workflow (e.g with github actions) to run the tests on each push and PR, possibly on different operating systems. Some coverage measure with codecov or coveralls would also be nice to have, although not compulsory.

CI workflow with github actions is added.
For local testing pytest is needed to be run on the tests folder.

> * [ ]  **Contributing**: I feel the current "How to contribute" section doesn't say how to contribute, as it merely encourages to read the documentation and browse the code. I would suggest some rewording, for example just replicating there the first two sentences of the Support section in the readme would make it much better. Style guide is missing, but fwiw I don't think this is strictly needed to be done before getting the paper published (but please do add that at some point :) )

Noted, the suggestion is adopted (although short but good for now). Regarding the style guide, will be on top of the list of the future works for the repo.

> * [ ]  **Functionality**: I get the same error reported by the other reviewer, a couple of more comments
>
>   * The example command in the readme has probably a typo, I think should be `offline_mot/main.py` -> `offlinemot/main.py`, same for the other examples

I think now it should work, ( a big download will be triggered when this command is run)
Offlinemot.main.main()

>   * That command works only assuming you are in the right place, this should be specified. Again, I think ideally I should be able to install the package and run it from anywhere
>   * Can I use the package with other detection models? Do they have to be pytorch or can it be a tensorflow model? It would be nice to have an example about this.

Sadly, the file should be `.pth` in order to be loaded, so pytorch is a requirement. But if the user wants to train on tensorflow, he needs simply to convert the output trained model to pytorch format. The model structure however, should be only Yolo v4. I think this note is added to the docs now.
Another detection structure is mentioned (other than Yolo). But Yolo is actually very common for the detection tasks so it’s not a bad choice.
>
> This was the first round about the code, I'll now read the paper.

> @engyasin I just downloaded all the files in your repo and followed your instructions, and then the error appeared. I will try the new checkpoint file like you said. But I recommend that you create a download script at the beginning of your Readme.md. That will make the followers and the audience much easier to follow.

The big file was for the Yolo trained network, (due to the big size of the network). Now it's only downloadable when the example model is run for the first time. it is no longer on github

> and here the second part of my review for the paper
>
> * [ ]  **Authorship**: could you clarify the contribution of Awad Mukbil and Jörg Müller to the package development? Based on the list of commits, they don't seem to be contributors

My two co-authors didn’t directly write the code for the repo, but had an active project participation and influencing directing. Upon review of the Journal guideline, active projects participants are considered for an authorship as well, so the credit was assigned that way.
Due to the nature of this work, which required designing process of the method and in specific details about extracting trajectories, discovering shadows and fixing the view, my co-authors had a direct involvement in the form of discussions and guidance. Other colleagues who collaborated with us for the training were listed in the acknowledgment part of the paper because their part was mainly about training the detection model not the method itself.

> * [ ]  **State of the field**: The paper at the moment compares the presented method to other methods, since this is a software paper, I think the focus should be on comparing the implementation to other existing packages. Are there other packages / codes tackling this problem, how does offlineMOT compare to them? At the moment, the section tells me some approaches have been developed but I am missing pointers to if and where they are implemented.

Comparing with other methods while interesting, still has some difficulties in our case,
First, we need a dataset with ground truth data in order to show the errors for all the methods at the end. These datasets should be compatible with our condition (stationary top-view). The biggest dataset for the field of multi-object tracking (MOT: https://motchallenge.net/ ) doesn't has such videos. Other datasets for drones (https://cvgl.stanford.edu/projects/uav_data/) don’t have initial bounding boxes for detection, so the detection should be done with pretrained models. This detection will produce its own errors, making the comparison less accurate.

What we showed instead is a new approach that combine well-known methods to solve restricted case for multi-objects tracking.
The paper mentions two methods among the state-of-the-art. [Deep Sort](https://github.com/nwojke/deep_sort) and [Tracktor ](https://github.com/phil-bergmann/tracking_wo_bnw).

> * [ ]  **Example usage**: I find this section in the current form a little confusing, I don't think installation instructions are needed in the paper, they should be covered in the documentation / readme. Furthermore, the current section is briefly showing a command line to call a specific script and commenting the results. If some usage examples are to be included, I think they should be more focused on the interface of the software and not the result. For example, before the pseudocode it is mentioned that the package is implemented in a modular way and it's easy to turn off and/or tune different parts, instead of the pseudocode, I think it would be more useful to have a code snippet showing how to build each block (if this can be done with a few lines of code, if the example is longer, better to just refer to the documentation)

This section was meant to walk through a complete example from start to end. but you are right regarding the point about installation, it isn't needed so It is now deleted.
For the modular design, a config file can be shown in parts, while walking through the example. There's a doc page explaining the many parameters, but perhaps in the paper it can be more configuration oriented than example oriented.
Now in the paper, this section is replaced with **Parameters Tuning** , with a table showing the most influencing parameters along with their effect.

I hope I did cover all the points, thanks again.
