+++ offlinemot/tool/darknet2pytorch.py
@@ -265,7 +265,8 @@ class Darknet(nn.Module):
                 elif activation == 'mish':
                     model.add_module('mish{0}'.format(conv_id), Mish())
                 else:
-                    print("convalution havn't activate {}".format(activation))
+                    pass
+                    #print("convalution havn't activate {}".format(activation))
 
                 prev_filters = filters
                 out_filters.append(prev_filters)
