+++ .github/workflows/pytest.yml
@@ -22,6 +22,7 @@ jobs:
           python3 -m pip install --upgrade pip
           pip3 install -r requirements.txt
           python3 -m pip install --upgrade pytest
+          pip3 install pytest-cov
 
       - name: Run test suite
         run: |
