+++ setup.py
@@ -18,7 +18,7 @@ except:
 setup(
   name = 'offlinemot',         # How you named your package folder (MyLib)
   packages = find_packages(),   # Chose the same as "name"
-  version = '1.0.2',      # Start with a small number and increase it with every change you make
+  version = '1.0.3',      # Start with a small number and increase it with every change you make
   license='MIT',        # Chose a license from here: https://help.github.com/articles/licensing-a-repository
   description = short_description[0],   # Give a short description about your library
   long_description=long_description,
@@ -26,7 +26,7 @@ setup(
   author = 'Yasin Yousif',                   # Type in your name
   author_email = 'mrhidestev@hotmail.com',      # Type in your E-Mail
   url = 'https://github.com/engyasin/Offline_MOT',   # Provide either the link to your github or to your website
-  download_url = 'https://github.com/engyasin/Offline_MOT/archive/refs/tags/v1.0.2.tar.gz',   # I explain this later on
+  download_url = 'https://github.com/engyasin/Offline_MOT/archive/refs/tags/v1.0.3.tar.gz',   # I explain this later on
   keywords = ['Multi Objects Tracking', 'Detection', 'Traffic Data'],   # Keywords that define your package best
   install_requires=[            # I get to this in a second
           'numpy',
