+++ docs/html/_sources/generated/offlinemot.config.config.cwd.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.config.config.cwd
-============================
-
-.. currentmodule:: offlinemot.config
-
-.. autoattribute:: config.cwd
