+++ docs/html/_sources/generated/offlinemot.config.config.window_size.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.config.config.window\_size
-=====================================
-
-.. currentmodule:: offlinemot.config
-
-.. autoattribute:: config.window_size
