+++ docs/html/_sources/generated/offlinemot.config.config.draw.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.config.config.draw
-=============================
-
-.. currentmodule:: offlinemot.config
-
-.. autoattribute:: config.draw
