+++ docs/html/_sources/generated/offlinemot.config.config.classes_file_name.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.config.config.classes\_file\_name
-============================================
-
-.. currentmodule:: offlinemot.config
-
-.. autoattribute:: config.classes_file_name
