+++ docs/html/_sources/installing.rst.txt
@@ -4,7 +4,7 @@ Installing
 
 The package can be installed on python 3.x simply using the ``pip`` command:
 
-.. code-block:: python
+.. code-block:: bash
 
     pip install offlinemot
 
@@ -16,9 +16,8 @@ After installing the library, and in order to test the example provided with it,
 
 .. code-block:: python
 
-    In [1]: import offlinemot
-
-    In [2]: offlinemot.main.main() # no input to run the example video
+    import offlinemot
+    offlinemot.core.extract_paths() # no input to run the example video
 
 
 In the first time this is ran, the example network model will be downloaded (around 250MB). And a window for the example video with the tracked objects will be shown.
@@ -35,24 +34,26 @@ The tracked objects will be surrounded with boxes in 5 different colors. Each co
     <li> <span style="color:gray">White</span>: The object is moving but still not classified to a class.</li>
     </ul>
 
-Of course, for a different case, the colors can be changed from the ``config.py`` file_. This also depends on the number of classes to predict.
+Of course, for a different case, the colors can be changed from the ``configs`` class instance here_. This also depends on the number of classes to predict.
 
 To control these parameters and many others, the commands to be run are:
 
 .. code-block:: python
 
-    offlinemot.main.set_params()
+    cfg = offlinemot.config.configs()
+     #input can be named file, or empty for the default values.
+    cfg['colors_map'] = [(255,0,0),(0,0,255),(0,255,0)]
 
 
 This will open a new text editor with the ``config.py`` file containing all of the parameters.
 
 .. note:: 
     It's highly recommended to set all the parameters when running on a new video. 
-    A detailed description for their meaning is available in the ``config`` file_ . 
+    A detailed description for their meaning is available in the ``config`` here_ . 
     Additionally, a complete example for parameters tuning is available here link_ 
 
 .. _link: tutorials/A_Working_Example.ipynb
-.. _file: generated/offlinemot.config.configs.html
+.. _here: generated/offlinemot.config.configs.html
 
 Running
 ----------
@@ -61,19 +62,20 @@ To run it on some video, the command is:
 
 .. code-block:: python
 
-    offlinemot.main.main('path_to_video') 
+    cfg = offlinemot.config.configs()
+    offlinemot.core.extract_paths('path_to_video',config=cfg) 
     #[directory of the videos, leave empty to run the example video]
 
 To show the result on the same video after post processing, use the command:
 
 .. code-block:: python
 
-    offlinemot.show_results.main('path_to_same_video')
+    offlinemot.show_results.show_result('path_to_same_video')
     #[directory of the videos, leave empty to run the example video]
 
 
-Finally, to change the yolo network used in the package, the complete directory to 3 files need to be assigned inside the ``config.py``:
+Finally, to change the yolo network used in the package, the complete directory to 3 files need to be assigned with the ``configs`` class instance:
 
-- *.pth* for the model weights
-- *.cfg* for the Yolo configuration.
-- *.names* for a simple text file containing the names of the classes.
+- *.pth* for the model weights  `cfg['model_name'] = 'directory'`
+- *.cfg* for the Yolo configuration. `cfg['model_config'] = 'directory'`
+- *.names* for a simple text file containing the names of the classes.  `cfg['class_file_name'] = 'directory'`
