+++ docs/html/_sources/generated/offlinemot.config.config.detect_every_N.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.config.config.detect\_every\_N
-=========================================
-
-.. currentmodule:: offlinemot.config
-
-.. autoattribute:: config.detect_every_N
