+++ docs/html/_sources/tutorials/Fixing_the_view.ipynb.txt
@@ -10,7 +10,7 @@
     "\n",
     "In the following example of the sample video all the frames should have the same position. \n",
     "\n",
-    "All the parameters for view fixing are in `config.py` file."
+    "All the parameters for view fixing are in `configs` class, section `Fix view`."
    ]
   },
   {
@@ -22,18 +22,18 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": 1,
    "metadata": {},
    "outputs": [],
    "source": [
     "from offlinemot.utils_ import resize\n",
     "from offlinemot.fix_view import FixView, BG_subtractor\n",
-    "from offlinemot.config import config"
+    "from offlinemot.config import configs"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 1,
+   "execution_count": 2,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -46,7 +46,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "Now the sample video should be read with a video reader object from Opencv liberary"
+    "Now the sample video should be read with a video reader object from Opencv library"
    ]
   },
   {
@@ -55,7 +55,8 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "cap = cv2.VideoCapture(os.path.join(config.cwd,'model','sample.mp4'))\n",
+    "cfg = configs()\n",
+    "cap = cv2.VideoCapture(os.path.join(cfg.cwd,'model','sample.mp4'))\n",
     "ret, bg = cap.read() # read the first frame"
    ]
   },
@@ -123,8 +124,8 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "Fix_obj = FixView(bg_rgb)\n",
-    "BG_s = BG_subtractor(bg)\n",
+    "Fix_obj = FixView(bg_rgb,config=cfg)\n",
+    "BG_s = BG_subtractor(bg,config=cfg)\n",
[…]
