+++ docs/html/_sources/generated/offlinemot.config.config.overlap_thresh.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.config.config.overlap\_thresh
-========================================
-
-.. currentmodule:: offlinemot.config
-
-.. autoattribute:: config.overlap_thresh
