+++ docs/html/_sources/generated/offlinemot.config.config.model_config.rst.txt
@@ -1,8 +1,0 @@
-:orphan:
-
-offlinemot.config.config.model\_config
-======================================
-
-.. currentmodule:: offlinemot.config
-
-.. autoattribute:: config.model_config

[…]
