### 2022-04-16 #4099 @engyasin
Dear editor @hugoledoux and reviewers @Shiming94 @lucaferranti ,

Some changes, according to the latest comments from @lucaferranti were made. Namely, the `configs` is a class now, providing Configparser interface of changing and showing the different parameters. So this will avoid the editing of a single file, and encourage saving different set of parameters to each case (and distributing it).

Also, *code coverage* is added. It shows only 51% however, many of the non-tested scripts are for creating the structure of Yolo network, and this mostly shouldn't be changed at all (all the scripts are in `tools` folder).

Contributing instructions are more detailed now. Instructions for creating a special type of issues for accuracy problem are shown as well.

Typos are fixed in the docs and code.

I hope I addressed your points. Of course, the improvement road is still wide open, but I hope a continuous cycle of enhancing would still be ongoing for the benefit of the community.

Thanks again,
