### 2022-04-19 #4099 @Shiming94
Hi @engyasin and @hugoledoux

Sorry for my delayed reply. I just moved to another country last month, and it was a really busy month.

I checked your repo again, and it is really great that the structure and contents of the code were significantly improved! But I was still stuck in the **Installation** step. I tried to run the code in a cond virtuel environment. The `pypi` installation works fine for me, but I got errors when I ran the first line in the `Getting Started`, here is the traceback:
 ```
/tmp/ipykernel_26571/510862866.py in <module>
----> 1 import offlinemot

~/12.Offline_MOT/offlinemot/__init__.py in <module>
     16 import fix_view
     17
---> 18 import objects_classes
     19
     20 import post_process

~/12.Offline_MOT/offlinemot/objects_classes.py in <module>
      7 from utils_ import check_box
      8
----> 9 class TrafficObj():
     10     """
     11     A class used to represent any traffic entity, with its full track

~/12.Offline_MOT/offlinemot/objects_classes.py in TrafficObj()
    106     """
    107
--> 108     def __init__(self,frame,frame_id,box,track_id,config=configs(),tracker=cv2.TrackerKCF_create,class_id=-1,detection_way=1,detect_prob=0.0):
    109         """
    110         Parameters

AttributeError: module 'cv2' has no attribute 'TrackerKCF_create'
```
I tried to tackle this problem by adapting the version of `opencv-python`, but they all failed. So there are several suggestions from my side:
- it seems that either the pypi installation or `requirements.txt` file can cover all the necessary dependencies or libraries. For example, I installed the jupyter notebook and opencv by myself. It would be really handy for some users if you could include the conda-env building steps with all the necessary dependencies and their versions in the `Installation` part.
-  It would also be great if you could creat a jupyter notebook file which contains the steps in `Getting Started`, so users don't need to type or paste them just for testing purposes.

I hope these are helpful for you. I will still try to tackle the open-cv problem, I hope it could be automatically solved when you update the dependencies' versions.
