### 2022-04-19 #4099 @engyasin
Hi @Shiming94 ,

Thanks for reporting. I think ,for opencv error, this has to do with the needed installation is "opencv-contrib-python" not "opencv-python", you may need also to uninstall opencv-python as mentioned here (https://github.com/opencv/opencv/issues/15723#issuecomment-543063004). This wouldn't happen if requirements.txt is used only, as shown in readme (pypi should use it as well). Just simply run :

`python -m pip install opencv-contrib-python`

Regarding the you second point, the readme should be md file, but there's some notebooks for starting in : https://github.com/engyasin/Offline_MOT/blob/main/docs/source/tutorials/A_Working_Example.ipynb
