### 2022-04-20 #4099 @hugoledoux
> Thanks for reporting. I think ,for opencv error, this has to do with the needed installation is "opencv-contrib-python" not "opencv-python", you may need also to uninstall opencv-python as mentioned here ([opencv/opencv#15723 (comment)](https://github.com/opencv/opencv/issues/15723#issuecomment-543063004)). This wouldn't happen if requirements.txt is used only, as shown in readme (pypi should use it as well). Just simply run :
> `pip uninstall opencv-python`
> `python -m pip install opencv-contrib-python`

@engyasin Maybe it would be good to stress this in the installation manual? I would assume this is something others might run into, and have it documented somewhere would be good practice.
