### 2022-04-20 #4099 @hugoledoux
> Regarding your second point, the readme should be md file, but there's some notebooks for starting in : https://github.com/engyasin/Offline_MOT/blob/main/docs/source/tutorials/A_Working_Example.ipynb

@engyasin Same here: inform the users in the readme.md that this exists and link to it?
