### 2022-04-20 #4099 @Shiming94
> Maybe it would be good to stress this in the installation manual? I would assume this is something others might run into, and have it documented somewhere would be good practice.

Hi @engyasin, thanks for your reply.

I tried the way you mentioned but it didn't work here. I think the problem is that we use different releases of opencv, and the API's name also changed. So my suggestion would be to give a clear description of which dependencies and which releases you are using before the `Installation`, so that users can check their environment.

Regarding the second part, I totally agree with @hugoledoux. You should let the users know that you already have these notebooks.
