### 2022-04-20 #4099 @engyasin
Hi @Shiming94 ,

Can you open an issue in the repo (showing the steps to get the error) , so we can focus on this. I couldn't reproduce the error. If the readme is followed, (just pip install on clean environment) everything worked for my case.

Secondly, I could direct the reader to the notebooks files. It could be easier than copy pasting for interested users. Noted.
