### 2022-04-22 #4099 @engyasin
> Hi @Shiming94 ,
>
> Can you open an issue in the repo (showing the steps to get the error) , so we can focus on this. I couldn't reproduce the error. If the readme is followed, (just pip install on clean environment) everything worked for my case.
>
> Secondly, I could direct the reader to the notebooks files. It could be easier than copy pasting for interested users. Noted.
>
> P.S: perhaps an upgrade is needed, to get the lateste fixes, `pip install -U offlinemot`

The test related to `Tracker Object` (https://github.com/engyasin/Offline_MOT/blob/fc4f1bb2196963ee1862e403745b24eedc93954a/offlinemot/tests/test_objects_classes.py#L18) also passed in Github actions.
