### 2022-05-10 #4099 @engyasin
Dear editor @hugoledoux and reviewers @Shiming94 @lucaferranti ,

Following @hugoledoux recommendation, I added the notebooks link to the readme.

**Please let me know if something else is needed.**
