+++ .github/workflows/pytest.yml
@@ -27,10 +27,10 @@ jobs:
       - name: Run test suite
         run: |
           pytest ./offlinemot/tests --exitfirst --verbose --failed-first \
-          --cov=./ --cov-report=xml
+          --cov=./  --cov-report=xml
 
       - name: "Upload coverage to Codecov"
         uses: codecov/codecov-action@v2
         with:
           fail_ci_if_error: true
-          token: ${{ secrets.CODECOV_TOKEN }} # not required for public repos
+          #token: ${{ secrets.CODECOV_TOKEN }} # not required for public repos
