+++ offlinemot/model/sample.mp4.txt
@@ -1,0 +1,19 @@
+3 [1752, 509, 62, 44] 1 0 166
+4 [1754, 508, 62, 44] 1 0 164
+5 [1753, 508, 62, 44] 1 0 165
+6 [1751, 509, 62, 44] 1 0 166
+7 [1747, 510, 62, 44] 1 0 170
+8 [1743, 511, 62, 44] 1 0 -177
+9 [1741, 511, 62, 44] 1 0 -177
+10 [1740, 512, 62, 44] 1 0 -177
+11 [1738, 513, 62, 44] 1 0 -177
+12 [1735, 513, 62, 44] 1 0 -177
+13 [1733, 514, 62, 44] 1 0 -177
+14 [1732, 514, 62, 44] 1 0 -177
+15 [1732, 513, 62, 44] 1 0 -177
+16 [1732, 513, 62, 44] 1 0 -177
+17 [1731, 513, 62, 44] 1 0 -177
+18 [1729, 510, 62, 44] 1 0 -177
+3 [1752, 509, 62, 44] 2 0 166
+4 [1754, 508, 62, 44] 2 0 164
+5 [1753, 508, 62, 44] 2 0 165
