### 2022-05-29 #4099 @engyasin
Dear reviewer @Shiming94 ,

Thanks,

1. I ran the command `pip install -e .`on windows, and it's working. It just links the local code to the python installation and it's indeed useful. I'm not aware of other modifications regarding this.
2. The package for training the network (that could be changed, i.e retrained) is YOLOv4 with this library (https://github.com/Tianxiaomo/pytorch-YOLOv4) which is mentioned both in the readme and paper.  Regarding the example, there are huge literature on training YOLO. Indeed, another packages can be used, like Darknet (https://github.com/AlexeyAB/darknet) and the result trained file can be converted to '.pth' format and used. I think this is mentioned also in the docs. Useful scripts are found also in these library mentioned in the docs.

> Hi @engyasin ,
>
> I basically finished the review. I successfully ran through all your examples and tutorials on both my laptop and workstation. Overall, it is very nice work, and you have improved a lot according to the recommendations of @hugoledoux and @lucaferranti. There are still several recommendations before publication:
>
> 1. Besides `PyPI` installation, it would be great if you could provide another installation method from source, like `pip install -e .`, which could be convenient for developers to further improve the repo.
> 2. I value the extensibility of the framework. You mentioned that other pretrained networks or backbones can also be possible. Can you provide some examples which could be used for this framework in the tutorials? If the model is separately trained, it would be great if the training script is also provided.
>
> Best, Shiming
