### 2022-05-29 #4099 @Shiming94
> 1. rking. It just links the local code to the python installation and it's indeed useful. I'm not aware of other modifications regarding this.

Dear @engyasin, Thanks for your reply. For the first point, you can just add this kind of installation method besides the ``PyPI` installation in the `Readme.md`. And the second point is pretty clear.

Dear editor @hugoledoux, I think the paper and the repo are also qualified for publication.

Best,
Shiming
