### 2022-06-09 12dc223
Some edits: fixes of typos, improvements to the text
