### 2022-06-09 #4099 @hugoledoux
@engyasin I went over the paper and fixed a few things (please use a spell-checker next time): https://github.com/engyasin/Offline_MOT/pull/1
