### 2022-06-14 #4099 @hugoledoux
> Version Tag: v1.0.9 [DOI: 10.5281/zenodo.6641975](https://doi.org/10.5281/zenodo.6641975)

That DOI points to v1.0.7 ?!

(last thing to fix, then we're good)
