### 2022-06-15 #4099 @arfon
@engyasin – I've found a couple of minor typos in the paper. Could you merge this PR please? https://github.com/engyasin/Offline_MOT/pull/2
