### 2022-06-15 #4099 @engyasin
> @engyasin – I've found a couple of minor typos in the paper. Could you merge this PR please? [engyasin/Offline_MOT#2](https://github.com/engyasin/Offline_MOT/pull/2)

Merged
